move common to common folder
