+++ imaging/imageAnonymization.ts
@@ -3,7 +3,7 @@
  */
 
 // internal libraries
-import { logger } from "../logger";
+import { logger } from "../common/logger";
 import { Series } from "./types";
 
 /*
