+++ imaging/imageManagers.ts
@@ -21,7 +21,7 @@ import type {
   FileManager
 } from "./types";
 import { getFileCustomImageId } from "./loaders/fileLoader";
-import { logger } from "../logger";
+import { logger } from "../common/logger";
 
 // global variables
 var imageManager: ImageManager = null;
@@ -251,12 +251,12 @@ export const populateGSPSManager = function (
         if (instanceUID) {
           gspsManager[instanceUID]
             ? gspsManager[instanceUID]!.push({
-                seriesId: prUniqueUID,
-                imageId: imageId
-              })
+              seriesId: prUniqueUID,
+              imageId: imageId
+            })
             : (gspsManager[instanceUID] = [
-                { seriesId: prUniqueUID, imageId: imageId }
-              ]);
+              { seriesId: prUniqueUID, imageId: imageId }
+            ]);
         }
       });
     }
