+++ imaging/imageParsing.ts
@@ -8,7 +8,7 @@ import { forEach, each, has, pick } from "lodash";
 import { v4 as uuidv4 } from "uuid";
 
 // internal libraries
-import { logger } from "../logger";
+import { logger } from "../common/logger";
 import { randomId } from "./imageUtils";
 import { getNestedObject, parseTag } from "./imageTags";
 import { updateLoadedStack } from "./imageLoading";
@@ -404,13 +404,13 @@ const fillMetadataReadable = function (metadata: MetaData): MetaDataReadable {
   const pixelSpacing = metadata.x00280030
     ? metadata.x00280030
     : metadata.x00080060 === "US" &&
-        metadata["x00186011"] != undefined &&
-        metadata["x00186011"][0].x0018602e != undefined &&
-        metadata["x00186011"][0].x0018602c != undefined
+      metadata["x00186011"] != undefined &&
+      metadata["x00186011"][0].x0018602e != undefined &&
+      metadata["x00186011"][0].x0018602c != undefined
       ? ([
-          metadata["x00186011"][0].x0018602e * 10, //so that from cm goes to mm
-          metadata["x00186011"][0].x0018602c * 10
-        ] as [number, number])
+        metadata["x00186011"][0].x0018602e * 10, //so that from cm goes to mm
+        metadata["x00186011"][0].x0018602c * 10
+      ] as [number, number])
       : metadata.x00181164
         ? metadata.x00181164
         : [1, 1];
@@ -430,7 +430,7 @@ const fillMetadataReadable = function (metadata: MetaData): MetaDataReadable {
   const numberOfTemporalPositions = metadata["x00200105"]; // Total number of temporal positions prescribed.
   const is4D =
     temporalPositionIdentifier !== undefined &&
-    (numberOfTemporalPositions as number) > 1
+      (numberOfTemporalPositions as number) > 1
       ? true
       : false;
