+++ imaging/loaders/dicomLoader.ts
@@ -7,7 +7,7 @@
 import cornerstone from "cornerstone-core";
 
 // internal libraries
-import { logger } from "../../logger";
+import { logger } from "../../common/logger";
 import { loadAndCacheImage, loadAndCacheImages } from "../imageRendering";
 import store, { set as setStore } from "../imageStore";
 import type { Series, CachingResponse } from "../types";
