+++ imaging/loaders/resliceLoader.ts
@@ -8,7 +8,7 @@ import cornerstone from "cornerstone-core";
 import { default as cornerstoneDICOMImageLoader } from "cornerstone-wado-image-loader";
 
 // internal libraries
-import { logger } from "../../logger";
+import { logger } from "../../common/logger";
 import { getImageFrame } from "./commonLoader";
 import { getImageTracker, getImageManager } from "../imageManagers";
 import type { Image, ImageFrame, MetaData } from "../types";
@@ -66,12 +66,12 @@ let createCustomImage = function (
   let pixelSpacing = metadata.x00280030
     ? metadata.x00280030
     : metadata.x00080060 === "US" &&
-        metadata["x00186011"]![0].x0018602e != undefined &&
-        metadata["x00186011"]![0].x0018602c != undefined
+      metadata["x00186011"]![0].x0018602e != undefined &&
+      metadata["x00186011"]![0].x0018602c != undefined
       ? ([
-          metadata["x00186011"]![0].x0018602e * 10, //so that from cm goes to mm
-          metadata["x00186011"]![0].x0018602c * 10
-        ] as [number, number])
+        metadata["x00186011"]![0].x0018602e * 10, //so that from cm goes to mm
+        metadata["x00186011"]![0].x0018602c * 10
+      ] as [number, number])
       : metadata.x00181164
         ? metadata.x00181164
         : [1, 1];
@@ -192,7 +192,7 @@ let createCustomImage = function (
     } else {
       logger.error(
         "Unable to calculate default window width/center for imageId: " +
-          imageId
+        imageId
       );
     }
   }
