+++ imaging/monitors/memory.ts
@@ -3,7 +3,7 @@
  *        monitoring memory usage
  */
 
-import { logger } from "../../logger";
+import { logger } from "../../common/logger";
 import { getImageManager, removeDataFromImageManager } from "../imageManagers";
 import { clearImageCache } from "../imageRendering";
 import store from "../imageStore";
