+++ imaging/tools/custom/gspsUtils/genericMathUtils.ts
@@ -1,5 +1,5 @@
 import { MetaData } from "../../../types";
-import type { ViewportComplete, Coords } from "../../types";
+import type { ViewportComplete, Coords } from "../../../../common/types";
 import cornerstone, {
   Image,
   PixelCoordinate,
